% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoch_gp.R
\name{smoch.gp}
\alias{smoch.gp}
\title{Synthetic Minority Ovsersampling Convex Hull - Gower's distance with Podani modification}
\usage{
smoch.gp(y, x, data, k = 5, oversampling = 100, outlier = FALSE, out_amp = 1.5)
}
\arguments{
\item{y}{A categorical variable with the unbalanced data.}

\item{x}{The variables we want to include in the prediction problem.}

\item{data}{A data frame containing the original (unbalanced) data set.}

\item{k}{A number indicating the number of nearest neighbours that are used to generate the new examples of the minority class.}

\item{oversampling}{A number that drives the decision of how many extra cases from the minority class are generated.}

\item{outlier}{Optionally, this parameter allows the method to be adapted for the presence of outliers.}

\item{out_amp}{A parameter to adjust the stringency of outlier detection.}
}
\value{
The function returns a list with three objects:

\bold{Distances} - The matrix of distances calculated between observations.

\bold{Nearest neighbors} - The nearest neighbors identified.

\bold{Newdata} - The newly generated balanced dataset.
}
\description{
Synthetic Minority Ovsersampling Convex Hull - Gower's distance with Podani modification
}
\examples{
df <- data.frame(y=rep(as.factor(c('Yes', 'No')), times=c(90, 10)), x1=rnorm(100), x2=rnorm(100))
smoch.gp(y='y', x=c('x1','x2'), data=df, k=5, oversampling = 100, outlier = F)
}
